requirements.txt
Flask==2.0.2
python-binance==1.0.15
pyTelegramBotAPI==4.2.2
python-dotenv==0.19.2
gunicorn==20.1.0
Werkzeug==2.0.3